requirements.txt
transformers
torch
chromadb
tiktoken
llama-index
huggingface_hub
python-docx
regex
spacy
fastapi
uvicorn
python-multipart
typing-extensions
aiofiles
pydantic
streamlit
io
requests
pillow
